some fixes

+++ src/main/java/game/Board.java
@@ -52,16 +52,16 @@ public class Board {
         return result;
     }
 
-    public int getFitness() {
+    public long getFitness() {
         /*
         int winnerPoints = result * 100;
         int snakeLength = snake.getBody().size();
         return winnerPoints + snakeLength*50 + moveCounter;*/
         int snakeLength = snake.getBody().size();
         if (snakeLength < 10) {
-            return (int) (moveCounter*moveCounter * Math.pow(2, snakeLength));
+            return (long) (moveCounter*moveCounter * Math.pow(2, snakeLength));
         }
-        return (int) (moveCounter*moveCounter*Math.pow(2, 10) * (snakeLength-9));
+        return (long) (moveCounter*moveCounter*Math.pow(2, 10) * (snakeLength-9));
     }
 
     public boolean isGameFinished() {
